refactor: remove silex.sile dependency
The only reasons why it was here is for the scalebox command
(introduced in SILE 0.14.9) and its re-implementation in silex.sile
for full-bleed printing. When used outside the resilient collection
this is not needed, and when used with it, the silex layers are
already loaded. So this change should not impact users.

Also, we cannot support forever older versions of SILE, so from
this point the minimal supported version is 0.14.13.